% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{getGene}
\alias{getGene}
\title{getGene}
\usage{
getGene(gene, filter = NULL)
}
\arguments{
\item{gene}{a character vector of HGNC gene symbols
A Function to get gene information from gene panels}
}
\description{
getGene
}
