% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{getPanel}
\alias{getPanel}
\title{getPanel}
\usage{
getPanel(panel, filter = NULL)
}
\arguments{
\item{panel}{a character vector of panel names or panel ids
A function to get gene panels}
}
\value{
A dataframe of gene panel and their genes
}
\description{
A function to query gene panels by name or id according to a
specified filter
}
\examples{
res <- getPanel(panel = '553f9595bb5a1616e5ed45a8')
res <- getPanel(panel = '553f9595bb5a1616e5ed45a8',
filter = list(ModeOfInheritance="monoallelic,biallelic"))
}
