% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.r
\name{paAlias2Name}
\alias{paAlias2Name}
\title{Convert to current panel name}
\usage{
paAlias2Name(ids, pad)
}
\arguments{
\item{ids}{character vector of panel names or aliases.}

\item{pad}{PanelApp data exported using function paDownload.}
}
\value{
vector with the current version of the panel name.
}
\description{
The function takes a character vector with panels or "aliases" names,
and returns the current name of the panel.
}
\details{
Information about the conversion is stored in pad$disorders
}
