% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paAnnotate.r
\name{paAnnotate}
\alias{paAnnotate}
\title{Annotate PanelApp information}
\usage{
paAnnotate(pad, append = TRUE, latestVersion = TRUE, verbose = TRUE)
}
\arguments{
\item{pad}{PanelApp data exported using function paDownload.}

\item{append}{if TRUE annotation information is included in new columns of pad$ensembl.
If FALSE a data frame with the collected information is returned.}

\item{latestVersion}{if TRUE current ensembl version is used.
If FALSE, ensembl version 80 is used.}

\item{verbose}{verbose.}
}
\value{
a pad element or a data.frame
}
\description{
The function annotates the "ensembl" data.frame in panelapp list created using paDownload.
}
\details{
Data are retrieved from Ensembl Biomart using the `biomaRt` library in Bioconductor.

Original gene ids in PanelApp defined using Ensembl 80 version.
Use latestVersion = FALSE if you want to retrieve the information from Ensembl 80.
Otherwise current ensembl version will be used.
}
\seealso{
paDownload
}
