% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paExport.r
\name{paExport.sqlite}
\alias{paExport.sqlite}
\title{Export PanelApp information to SQLite}
\usage{
paExport.sqlite(pad, file, verbose = TRUE, protect = TRUE)
}
\arguments{
\item{pad}{PanelApp data exported using function paDownload.}

\item{file}{SQLite file (the file extension has to be included).}

\item{verbose}{verbose.}

\item{protect}{when FALSE the file is overwritten if it exists.}
}
\value{
an SQLite file.
}
\description{
The function saves PanelApp data into a SQLite database file.
}
