% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paExport.r
\name{paExport.xlsx}
\alias{paExport.xlsx}
\title{Export PanelApp information as xlsx file}
\usage{
paExport.xlsx(pad, file)
}
\arguments{
\item{pad}{PanelApp data exported using function paDownload.}

\item{file}{xls or xlsx file name (the file extension has to be included).}
}
\value{
an xmlx file.
}
\description{
The function saves PanelApp data into an xlsx file using the "openxlsx" library.
}
\details{
Tables "publications" and "phenotypes" are NOT into the xlsx format
because they contain end of lines and tabs which break the xlsx file.
}
