% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paGenPanMat.r
\name{paGenPanMat}
\alias{paGenPanMat}
\title{Gene to Panel indicator matrix}
\usage{
paGenPanMat(pad, logical = TRUE)
}
\arguments{
\item{pad}{PanelApp data exported using function paDownload.}

\item{logical}{if TRUE a logical matrix is returned.
If FALSE a binary matrix (0, 1) is created.}
}
\value{
A matrix. Genes arranged in rows and panels in columns.
}
\description{
The function creates a gene to panel indicator matrix from PanelApp
information downloaded using paDownload.
}
