% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.r
\name{paName2CurrentVersion}
\alias{paName2CurrentVersion}
\title{Find current panel version}
\usage{
paName2CurrentVersion(ids, pad)
}
\arguments{
\item{ids}{character vector of panel names.}

\item{pad}{PanelApp data exported using function paDownload.}
}
\value{
vector with the current version number of the panel.
}
\description{
The function takes a character vector with panel names
and returns the current (latest) version of the panel.
}
\details{
Information about the conversion is stored in pad$panels$CurrentVersion.
The function just works with latest panel names.
Use `paAlias2Name` if you need to convert aliases.
}
