% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paNeedsTrim.r
\name{paNeedsTrim}
\alias{paNeedsTrim}
\title{Find data which need trimming}
\usage{
paNeedsTrim(pad)
}
\arguments{
\item{pad}{PanelApp data exported using funciton paDownload.}
}
\value{
a list of data frames
}
\description{
The function creates a list of data frames keeping the rows which need some trimming.
}

