% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_query.R
\name{getGene}
\alias{getGene}
\title{getGene}
\usage{
getGene(gene, filter = NULL)
}
\arguments{
\item{gene}{a character vector of HGNC gene symbols}
}
\value{
A dataframe of gene panels for the specified gene/s
}
\description{
A function to qurey gene panels by gene#'
}
\examples{
res <- getGene("BRCA1")
res <- getGene("BRCA1", filter=list(ModeOfPathogenicity="loss_of_function"))

}
