% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paDownload.r
\name{paDownload}
\alias{paDownload}
\title{Download PanelApp information}
\usage{
paDownload(verbose = TRUE)
}
\arguments{
\item{verbose}{verbose.}
}
\value{
a list of data.frames with the following information

- panels
- disorders
- genes
- evidences
- publications
- phenotypes
- ensembl
}
\description{
The function downloads PanelApp latest version  information
}
