% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paNeedsTrim.r
\name{paNeedsTrim}
\alias{paNeedsTrim}
\title{Find rows which need trimming}
\usage{
paNeedsTrim(pad)
}
\arguments{
\item{pad}{PanelApp data exported using funciton paDownload.}
}
\value{
A list of data frames with the same items as in the `pad` and filtered rows.
}
\description{
The function filters the rows in the data frames downloaded from Panelapp.
For each data frame, it keeps just those rows which need white trimming
in any of their columns.
}
\details{
The columns in the data frames in the out put list are all converted to text.
Moreover, all values in the output are wrapped withing double quotes
so that the white spaces are easy to spot.

By default "needs trimming" means that whites apear at the beggining
and at the end of the character values.
}
